Refactor entity relationships: move `Entry` association from `Product` to `WarehouseProduct`, update methods, and adjust UML diagram accordingly.
